// src/app/core/models/reponses.ts
export class Reponses {
    id: number | undefined;
    idReponse: string | undefined;
    idQuestionnaire: any | undefined;
    idQuestion: number | undefined;
    description: string | undefined;
    isQuestion: boolean | undefined;
    numeroOrdre: number | undefined;
    dateDebut: String | undefined;
    dateFin: String | undefined;
    auditUser: string| undefined;
  questionnaire: any;
}